unified_mode and custom resource for mdadm
